backend/requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.10
astroid==2.5.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.5
black==20.8b1
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
dataclasses==0.6
defusedxml==0.6.0
Django==2.2
django-allauth==0.22.0
django-cors-headers==3.4.0
django-crispy-forms==1.10.0
django-easy-pdf==0.1.1
django-email-verification==0.1.0
django-guardian==2.3.0
django-jsonfield==1.4.0
django-rest-passwordreset==1.1.0
django-six==1.0.4
django-taggit==1.2.0
django-truncate==0.1
django-xhtml2pdf==0.0.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
environ==1.0
future==0.18.2
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.1
gunicorn==20.0.4
h11==0.11.0
h2==4.0.0
hpack==4.0.0
html5lib==1.1
Hypercorn==0.11.1
hyperframe==6.0.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
pathspec==0.8.1
pep8==1.7.1
Pillow==7.2.0
priority==1.3.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.7.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
Quart==0.13.1
regex==2020.11.13
reportlab==3.5.59
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==3.3.1
wrapt==1.12.1
wsproto==0.15.0
xhtml2pdf==0.2.5
zope.interface==5.1.2
